+++ post/index.html
@@ -60,7 +60,7 @@
         </div>
       </div>
     </header>
-    <form id="create-post-form">
+    <form id="postForm">
       <div class="container mt-5">
         <div class="form-group">
           <label for="exampleInputEmail1">Title here</label>
