+++ profil/login/index.html
@@ -14,7 +14,7 @@
       src="https://kit.fontawesome.com/9e8ffdbe7c.js"
       crossorigin="anonymous"
     ></script>
-    <script src="/src/js/index.mjs" defer type="module"></script>
+    <script defer src="../../src/js/index.mjs"  type="module"></script>
   </head>
   <body>
     <img
