+++ src/js/handlers/login.mjs
@@ -1,9 +1,9 @@
 import { login } from "../api/auth/login.mjs";
 
 export function setloginFormListener() {
-  const form = document.querySelector("#loginForm");
+  const accessform = document.querySelector("#loginForm");
 
-  form.addEventListener("submit", (event) => {
+  accessform.addEventListener("submit", (event) => {
     event.preventDefault();
     const form = event.target;
     const formData = new FormData(form);
