+++ src/js/index.mjs
@@ -2,6 +2,7 @@
 import { setRegisterFormListener } from "./handlers/register.mjs";
 import { setloginFormListener } from "./handlers/login.mjs";
 import * as post from "./api/posts/index.mjs";
+import { renderPostToHtml } from "./handlers/renderpost.mjs";
 
 const path = location.pathname;
 
@@ -10,5 +11,3 @@ if (path === "/profil/login/index.html") {
 } else if (path === "/profil/register/index.html") {
   setRegisterFormListener();
 }
-
-post.createPost();
